challenge 2
You have to make changes only on the file ViewController.Swift
#Challenge:
Let users filter the petitions they see. This involves creating a second array of filtered items that contains only petitions matching a string the user entered. Use a UIAlertController with a text field to let them enter that string.